--- src/config/_config_old.js
@@ -1,8 +1,0 @@
-// module.exports = {
-//   port: process.env.PORT,
-//   db: process.env.DB,
-//   db_user: process.env.DB_USER,
-//   db_password: process.env.DB_PASSWORD,
-//   host: process.env.HOST,
-//   env: process.env.NODE_ENV,
-// };
