+++ src/features/GameContainer/GameContainer.tsx
@@ -14,9 +14,10 @@ import {
 import { shuffleLetters, calcWordScore, anagramsByLevelWord } from '../../utils/methods.util';
 import Button from '../../components/Button/Button';
 import AnagramService from '../../services/anagram.service';
-import {
-  AnagramsType, Anagram, GameStatus, GameContainerProps, Letter, NotificationKey,
-} from '../../utils/types.util';
+import { GameStatus, GameContainerProps } from '../../types/game.interface';
+import { NotificationKey } from '../../types/notification.interface';
+import { Letter } from '../../types/letter.interface';
+import { Anagram, AnagramsHashMap } from '../../types/anagram.interface';
 
 function GameContainer(props: GameContainerProps) {
   const {
@@ -36,7 +37,7 @@ function GameContainer(props: GameContainerProps) {
   });
 
   const [gameLetters, setGameLetters] = useState<Letter[]>([]);
-  const [anagrams, setAnagrams] = useState<AnagramsType>({});
+  const [anagrams, setAnagrams] = useState<AnagramsHashMap>({});
   const [notification, setNotification] = useState<NotificationKey>(defaultNotification);
 
   const { score, level, levelUp } = player;
