+++ src/types/anagram.interface.ts
@@ -1,3 +1,5 @@
+import { GameStatus } from './game.interface';
+
 export interface BaseAnagram {
   id: number,
   anagram: string,
@@ -6,4 +8,16 @@ export interface BaseAnagram {
 
 export interface Anagram extends BaseAnagram {
   solved: boolean
+}
+
+export interface AnagramsHashMap {
+  [key: string]: {
+    [key: string]: Anagram
+  }
+}
+
+export interface AnagramsProps {
+  gameStatus: GameStatus,
+  anagrams: AnagramsHashMap,
+  levelWordText: string,
 }
