+++ src/types/api.interface.ts
@@ -1,5 +1,8 @@
 import { Query } from 'express-serve-static-core';
 
+// Express types
 export interface TypedRequestQuery<T extends Query> extends Express.Request {
   query: T
 }
+
+export type ApiResponse<T> = Promise<T>;
