+++ src/types/misc.types.ts
@@ -1,0 +1,22 @@
+import {
+  scoreMultiples,
+} from '../utils/constants.util';
+
+export type ButtonProps = {
+  className?: string,
+  text?: string,
+  icon?: string
+};
+
+export type TextInputProps = {
+  className?: string,
+};
+
+export type GuessMobileProps = {
+  userGuess: string,
+  handleBackspace: () => void;
+};
+
+export type ShuffleLetters = () => ShuffleLetters | string;
+
+export type ScoreWordLengthKey = keyof typeof scoreMultiples;
