+++ src/types/start-page.interface.ts
@@ -1,0 +1,6 @@
+import { GameStatus } from './game.interface';
+
+export type StartPageProps = {
+  updateGameStatus: (status: GameStatus) => void,
+  hasLevelWord: boolean
+};
